+++ src/main/java/org/eclipse/keyple/card/calypso/crypto/legacysam/CommandGiveRandom.java
@@ -45,8 +45,8 @@ final class CommandGiveRandom extends Command {
     super(CommandRef.GIVE_RANDOM, 0, context);
 
     byte cla = context.getTargetSam().getClassByte();
-    byte p1 = (byte) 0x00;
-    byte p2 = (byte) 0x00;
+    byte p1 = 0x00;
+    byte p2 = 0x00;
 
     if (random == null || random.length != 8) {
       throw new IllegalArgumentException("Random value should be an 8 bytes long");