+++ src/main/java/org/eclipse/keyple/card/calypso/crypto/legacysam/DtoAdapters.java
@@ -63,7 +63,7 @@ final class DtoAdapters {
      * @since 0.1.0
      */
     @Override
-    public T setData(byte[] data, byte kif, byte kvc) {
+    public final T setData(byte[] data, byte kif, byte kvc) {
       this.data = data;
       this.kif = kif;
       this.kvc = kvc;
@@ -76,8 +76,8 @@ final class DtoAdapters {
      * @since 0.1.0
      */
     @Override
-    public T setSignatureSize(int size) {
-      this.signatureSize = size;
+    public final T setSignatureSize(int size) {
+      signatureSize = size;
       return currentInstance;
     }
 
@@ -87,8 +87,8 @@ final class DtoAdapters {
      * @since 0.1.0
      */
     @Override
-    public T setKeyDiversifier(byte[] diversifier) {
-      this.keyDiversifier = diversifier;
+    public final T setKeyDiversifier(byte[] diversifier) {
+      keyDiversifier = diversifier;
       return currentInstance;
     }
 
@@ -98,7 +98,7 @@ final class DtoAdapters {
      * @since 0.1.0
      */
     @Override
-    public byte[] getSignature() {
+    public final byte[] getSignature() {
       if (signature == null) {
         throw new IllegalStateException(MSG_THE_COMMAND_HAS_NOT_YET_BEEN_PROCESSED);
       }
@@ -109,7 +109,7 @@ final class DtoAdapters {
      * @return A not empty array of data. It is required to check input data first.
      * @since 0.1.0
      */
-    byte[] getData() {
+    final byte[] getData() {
       return data;
     }
 
@@ -117,7 +117,7 @@ final class DtoAdapters {
      * @return The KIF. It is required to check input data first.
      * @since 0.1.0
      */
-    byte getKif() {
+    final byte getKif() {
       return kif;
     }
 
@@ -125,7 +125,7 @@ final class DtoAdapters {
      * @return The KVC. It is required to check input data first.
      * @since 0.1.0
      */
-    byte getKvc() {
+    final byte getKvc() {
       return kvc;
     }
 
@@ -133,7 +133,7 @@ final class DtoAdapters {
      * @return The signature size.
      * @since 0.1.0
      */
-    int getSignatureSize() {
+    final int getSignatureSize() {
       return signatureSize;
     }
 
@@ -141,7 +141,7 @@ final class DtoAdapters {
      * @return Null if the key diversifier is not set.
      * @since 0.1.0
      */
-    byte[] getKeyDiversifier() {
+    final byte[] getKeyDiversifier() {
       return keyDiversifier;
     }
 
@@ -151,7 +151,7 @@ final class DtoAdapters {
      * @param signature The computed signature.
      * @since 0.1.0
      */
-    void setSignature(byte[] signature) {
+    final void setSignature(byte[] signature) {
       this.signature = signature;
     }
   }
@@ -180,7 +180,7 @@ final class DtoAdapters {
      * @since 0.1.0
      */
     @Override
-    public T setData(byte[] data, byte[] signature, byte kif, byte kvc) {
+    public final T setData(byte[] data, byte[] signature, byte kif, byte kvc) {
       this.data = data;
       this.signature = signature;
       this.kif = kif;
@@ -194,8 +194,8 @@ final class DtoAdapters {
      * @since 0.1.0
      */
     @Override
-    public T setKeyDiversifier(byte[] diversifier) {
-      this.keyDiversifier = diversifier;
+    public final T setKeyDiversifier(byte[] diversifier) {
+      keyDiversifier = diversifier;
       return currentInstance;
     }
 
@@ -205,7 +205,7 @@ final class DtoAdapters {
      * @since 0.1.0
      */
     @Override
-    public boolean isSignatureValid() {
+    public final boolean isSignatureValid() {
       if (isSignatureValid == null) {
         throw new IllegalStateException(MSG_THE_COMMAND_HAS_NOT_YET_BEEN_PROCESSED);
       }
@@ -216,7 +216,7 @@ final class DtoAdapters {
      * @return A not empty array of data. It is required to check input data first.
      * @since 0.1.0
      */
-    byte[] getData() {
+    final byte[] getData() {
       return data;
     }
 
@@ -225,7 +225,7 @@ final class DtoAdapters {
      *     first.
      * @since 0.1.0
      */
-    byte[] getSignature() {
+    final byte[] getSignature() {
       return signature;
     }
 
@@ -233,7 +233,7 @@ final class DtoAdapters {
      * @return The KIF. It is required to check input data first.
      * @since 0.1.0
      */
-    byte getKif() {
+    final byte getKif() {
       return kif;
     }
 
@@ -241,7 +241,7 @@ final class DtoAdapters {
      * @return The KVC. It is required to check input data first.
      * @since 0.1.0
      */
-    byte getKvc() {
+    final byte getKvc() {
       return kvc;
     }
 
@@ -249,7 +249,7 @@ final class DtoAdapters {
      * @return Null if the key diversifier is not set.
      * @since 0.1.0
      */
-    byte[] getKeyDiversifier() {
+    final byte[] getKeyDiversifier() {
       return keyDiversifier;
     }
 
@@ -259,7 +259,7 @@ final class DtoAdapters {
      * @param isSignatureValid True if the signature is valid.
      * @since 0.1.0
      */
-    void setSignatureValid(boolean isSignatureValid) {
+    final void setSignatureValid(boolean isSignatureValid) {
       this.isSignatureValid = isSignatureValid;
     }
   }
@@ -305,9 +305,9 @@ final class DtoAdapters {
     @Override
     public TraceableSignatureComputationData withSamTraceabilityMode(
         int offset, boolean usePartialSamSerialNumber) {
-      this.isSamTraceabilityMode = true;
-      this.traceabilityOffset = offset;
-      this.isPartialSamSerialNumber = usePartialSamSerialNumber;
+      isSamTraceabilityMode = true;
+      traceabilityOffset = offset;
+      isPartialSamSerialNumber = usePartialSamSerialNumber;
       return this;
     }
 
@@ -318,7 +318,7 @@ final class DtoAdapters {
      */
     @Override
     public TraceableSignatureComputationData withoutBusyMode() {
-      this.isBusyMode = false;
+      isBusyMode = false;
       return this;
     }
 
@@ -404,8 +404,8 @@ final class DtoAdapters {
     @Override
     public TraceableSignatureVerificationData withSamTraceabilityMode(
         int offset, boolean isPartialSamSerialNumber, LSRevocationServiceSpi samRevocationService) {
-      this.isSamTraceabilityMode = true;
-      this.traceabilityOffset = offset;
+      isSamTraceabilityMode = true;
+      traceabilityOffset = offset;
       this.isPartialSamSerialNumber = isPartialSamSerialNumber;
       this.samRevocationService = samRevocationService;
       return this;
@@ -418,7 +418,7 @@ final class DtoAdapters {
      */
     @Override
     public TraceableSignatureVerificationData withoutBusyMode() {
-      this.isBusyMode = false;
+      isBusyMode = false;
       return this;
     }
 
@@ -500,8 +500,8 @@ final class DtoAdapters {
      */
     ApduRequestAdapter(byte[] apdu) {
       this.apdu = apdu;
-      this.successfulStatusWords = new HashSet<Integer>();
-      this.successfulStatusWords.add(DEFAULT_SUCCESSFUL_CODE);
+      successfulStatusWords = new HashSet<Integer>();
+      successfulStatusWords.add(DEFAULT_SUCCESSFUL_CODE);
     }
 
     /**
@@ -514,7 +514,7 @@ final class DtoAdapters {
      * @since 0.1.0
      */
     ApduRequestAdapter addSuccessfulStatusWord(int successfulStatusWord) {
-      this.successfulStatusWords.add(successfulStatusWord);
+      successfulStatusWords.add(successfulStatusWord);
       return this;
     }
 
@@ -538,7 +538,7 @@ final class DtoAdapters {
      * @return The object instance.
      * @since 0.1.0
      */
-    ApduRequestAdapter setInfo(final String info) {
+    ApduRequestAdapter setInfo(String info) {
       this.info = info;
       return this;
     }
@@ -560,7 +560,7 @@ final class DtoAdapters {
      */
     @Override
     public byte[] getApdu() {
-      return this.apdu;
+      return apdu;
     }
 
     /**
@@ -644,12 +644,11 @@ final class DtoAdapters {
 
     private static final int DEFAULT_SUCCESSFUL_CODE = 0x9000;
 
-    private String cardProtocol;
     private String powerOnDataRegex;
     private final Set<Integer> successfulSelectionStatusWords;
 
     /**
-     * Created an instance of {@link CardSelectorAdapter}.
+     * Creates a new instance.
      *
      * <p>Initialize default values.
      *
@@ -682,8 +681,8 @@ final class DtoAdapters {
      * @since 0.1.0
      */
     @Override
-    public final String getCardProtocol() {
-      return cardProtocol;
+    public String getCardProtocol() {
+      return null;
     }
 
     /**
@@ -887,7 +886,7 @@ final class DtoAdapters {
    *
    * @since 0.3.0
    */
-  static class CommandContextDto {
+  static final class CommandContextDto {
 
     private final LegacySamAdapter targetSam;
     private final ProxyReaderApi controlSamReader;
@@ -901,7 +900,7 @@ final class DtoAdapters {
      * @param controlSam The control legacy SAM.
      * @since 0.3.0
      */
-    public CommandContextDto(
+    CommandContextDto(
         LegacySamAdapter targetSam, ProxyReaderApi controlSamReader, LegacySamAdapter controlSam) {
       this.targetSam = targetSam;
       this.controlSamReader = controlSamReader;